requirements.txt
# Core dependencies
torch>=1.9.0
transformers>=4.15.0
accelerate>=0.12.0

# API
fastapi>=0.68.0
uvicorn[standard]>=0.15.0
pydantic>=1.8.0

# MLOps & Monitoring
mlflow>=1.20.0
prometheus-client>=0.11.0
psutil>=5.8.0

# Data processing
numpy>=1.21.0
pandas>=1.3.0
pyyaml>=5.4.0
python-dotenv>=0.19.0

# Testing
pytest>=6.2.0
pytest-cov>=2.12.0

# Utils
python-multipart>=0.0.5
joblib>=1.0.0


torch>=2.0.0
transformers>=4.30.0
accelerate>=0.20.0
sentencepiece>=0.1.99
protobuf>=3.20.0